chore(web-app): rename page header large component
